Joining the backend functions and first PTI result!

+++ expressways/calculation/models.py
@@ -4,5 +4,5 @@ from django.contrib.postgres.fields import JSONField
 class CalculationResult(models.Model):
     task_id = models.CharField(max_length=100)
     items = JSONField()
-    objective_1 = models.PositiveIntegerField()
-    objective_2 = models.PositiveIntegerField()
+    objective_1 = models.DecimalField(max_digits=7, decimal_places=3)
+    objective_2 = models.DecimalField(max_digits=7, decimal_places=3)
